requirements.txt
pyserial
six
enum34
